parseVlog better identifies methods that remain interpreted
Before this commit, if a method M was compiled and then
the recompilation failed and it was not retried, the script
would identify method M as an interpreted method.
However, the JIT will not retry a recompilation failure
rather relying on the existing compiled body.
This commit fixes this problem.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>